requirements.txt
streamlit_lottie == 0.0.3
streamlit == 1.10.0
requests == 2.26.0
Pillow == 8.3.1
python == 3.8.8
pip == 22.1